+++ phlib/json.c
@@ -892,7 +892,7 @@ PPH_STRING PhGetXmlNodeAttributeText(
 
 PCSTR PhGetXmlNodeAttributeByIndex(
     _In_ PVOID XmlNodeObject,
-    _In_ LONG Index,
+    _In_ SIZE_T Index,
     _Out_ PCSTR* AttributeName
     )
 {
